Refactored sha1 & sha2 into different files
